Volume-2-Practical-Applications/Code/requirements.txt
# Volume 2: Advanced Applications - Requirements
# AI for Networking Engineers by Eduard Dulharu

# Core LLM APIs
anthropic>=0.18.0
openai>=1.12.0

# LangChain Ecosystem
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.20
langsmith>=0.0.70

# Vector Databases & Embeddings
chromadb>=0.4.22
faiss-cpu>=1.7.4

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
markdown>=3.5.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0

# Network Automation (optional)
netmiko>=4.3.0
paramiko>=3.4.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Development
ipython>=8.20.0
jupyter>=1.0.0
